+++ app/src/main/java/woowacourse/shopping/data/product/ProductDummy.kt
@@ -1,0 +1,214 @@
+package woowacourse.shopping.data.product
+
+import woowacourse.shopping.domain.product.Money
+import woowacourse.shopping.domain.product.Product
+
+object ProductDummy {
+    val products =
+        listOf(
+            Product(
+                id = 1L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20220215/2359611/2359611_17024446698621_big.jpg?w=1200",
+                name = "오버사이즈 칼라드 스웨트셔츠 [블랙]",
+                _price = Money(35_090),
+            ),
+            Product(
+                id = 2L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230202/3054403/3054403_17286357268519_big.jpg?w=1200",
+                name = "릴렉스드 린넨 라이크 셔츠 [네이비]",
+                _price = Money(33_890),
+            ),
+            Product(
+                id = 3L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20240219/3879193/3879193_17128786437292_big.jpg?w=1200",
+                name = "케이블 니티드 폴로 셔츠 [아이리시 그린]",
+                _price = Money(39_090),
+            ),
+            Product(
+                id = 4L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20241213/4664346/4664346_17405306803729_big.jpg?w=1200",
+                name = "린넨 라이크 미니멀 데님 트러커 재킷 [로우 블랙]",
+                _price = Money(47_890),
+            ),
+            Product(
+                id = 5L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230712/3404707/3404707_17134032325426_big.jpg?w=1200",
+                name = "시티 레저 와이드 벨티드 카고 팬츠 [블랙]",
+                _price = Money(35_990),
+            ),
+            Product(
+                id = 6L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230712/3404707/3404707_17134032325426_big.jpg?w=1200",
+                name = "시티 레저 와이드 벨티드 카고 팬츠 [블랙]",
+                _price = Money(35_990),
+            ),
+            Product(
+                id = 7L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190322/991339/991339_17138316859815_big.jpg?w=1200",
+                name = "라이트웨이트 크루 삭스 7팩 [화이트]",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 8L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20191112/1220731/1220731_17073700380718_big.jpg?w=1200",
+                name = "사피아노 신세틱 레더 벨트_30mm",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 9L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190910/1149329/1149329_16760172077751_big.jpg?w=1200",
+                name = "세미 와이드 히든 밴딩 슬랙스 [블랙]",
+                _price = Money(24_990),
+            ),
+            Product(
+                id = 10L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20210825/2086653/2086653_17048764754484_big.jpg?w=1200",
+                name = "레이어드 크루 넥 반팔 티셔츠_일반 기장 [화이트]",
+                _price = Money(14_790),
+            ),
+            Product(
+                id = 11L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20210719/2034137/2034137_17138591099209_big.jpg?w=1200",
+                name = "릴렉스 핏 크루 넥 반팔 티셔츠 2팩",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 12L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190322/991340/991340_17138316454543_big.jpg?w=1200",
+                name = "라이트웨이트 크루 삭스 7팩 [블랙]",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 13L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230201/3051675/3051675_16806757236998_big.jpg?w=1200",
+                name = "린넨 라이크 와이드 데님 팬츠 [로우 블랙]",
+                _price = Money(41_390),
+            ),
+            Product(
+                id = 14L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230712/3404705/3404705_17134029927665_big.jpg?w=1200",
+                name = "시티 레저 CN 와이드 벨티드 파라슈트 팬츠 [블랙]",
+                _price = Money(34_990),
+            ),
+            Product(
+                id = 15L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190327/996177/996177_17050276762608_big.jpg?w=1200",
+                name = "릴렉스 핏 크루 넥 반팔 티셔츠 [화이트]",
+                _price = Money(13_890),
+            ),
+            Product(
+                id = 16L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190712/1094748/1094748_17138345536257_big.jpg?w=1200",
+                name = "라이트웨이트 미들 삭Money(스 7팩 )[화이트]",
+                _price = Money(10_990),
+            ),
+            Product(
+                id = 17L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20210901/2101924/2101924_17138482498938_big.jpg?w=1200",
+                name = "플레인 더비 슈즈 [블랙]",
+                _price = Money(89_890),
+            ), // 기존 더미 마지막
+            Product(
+                id = 18L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20220215/2359611/2359611_17024446698621_big.jpg?w=1200",
+                name = "오버사이즈 칼라드 스웨트셔츠 [블랙]",
+                _price = Money(35_090),
+            ),
+            Product(
+                id = 19L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230202/3054403/3054403_17286357268519_big.jpg?w=1200",
+                name = "릴렉스드 린넨 라이크 셔츠 [네이비]",
+                _price = Money(33_890),
+            ),
+            Product(
+                id = 20L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20240219/3879193/3879193_17128786437292_big.jpg?w=1200",
+                name = "케이블 니티드 폴로 셔츠 [아이리시 그린]",
+                _price = Money(39_090),
+            ),
+            Product(
+                id = 21L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20241213/4664346/4664346_17405306803729_big.jpg?w=1200",
+                name = "린넨 라이크 미니멀 데님 트러커 재킷 [로우 블랙]",
+                _price = Money(47_890),
+            ),
+            Product(
+                id = 22L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230712/3404707/3404707_17134032325426_big.jpg?w=1200",
+                name = "시티 레저 와이드 벨티드 카고 팬츠 [블랙]",
+                _price = Money(35_990),
+            ),
+            Product(
+                id = 23L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230712/3404707/3404707_17134032325426_big.jpg?w=1200",
+                name = "시티 레저 와이드 벨티드 카고 팬츠 [블랙]",
+                _price = Money(35_990),
+            ),
+            Product(
+                id = 24L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190322/991339/991339_17138316859815_big.jpg?w=1200",
+                name = "라이트웨이트 크루 삭스 7팩 [화이트]",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 25L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20191112/1220731/1220731_17073700380718_big.jpg?w=1200",
+                name = "사피아노 신세틱 레더 벨트_30mm",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 26L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190910/1149329/1149329_16760172077751_big.jpg?w=1200",
+                name = "세미 와이드 히든 밴딩 슬랙스 [블랙]",
+                _price = Money(24_990),
+            ),
+            Product(
+                id = 27L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20210825/2086653/2086653_17048764754484_big.jpg?w=1200",
+                name = "레이어드 크루 넥 반팔 티셔츠_일반 기장 [화이트]",
+                _price = Money(14_790),
+            ),
+            Product(
+                id = 28L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20210719/2034137/2034137_17138591099209_big.jpg?w=1200",
+                name = "릴렉스 핏 크루 넥 반팔 티셔츠 2팩",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 29L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190322/991340/991340_17138316454543_big.jpg?w=1200",
+                name = "라이트웨이트 크루 삭스 7팩 [블랙]",
+                _price = Money(9_900),
+            ),
+            Product(
+                id = 30L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230201/3051675/3051675_16806757236998_big.jpg?w=1200",
+                name = "린넨 라이크 와이드 데님 팬츠 [로우 블랙]",
+                _price = Money(41_390),
+            ),
+            Product(
+                id = 31L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20230712/3404705/3404705_17134029927665_big.jpg?w=1200",
+                name = "시티 레저 CN 와이드 벨티드 파라슈트 팬츠 [블랙]",
+                _price = Money(34_990),
+            ),
+            Product(
+                id = 32L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190327/996177/996177_17050276762608_big.jpg?w=1200",
+                name = "릴렉스 핏 크루 넥 반팔 티셔츠 [화이트]",
+                _price = Money(13_890),
+            ),
+            Product(
+                id = 33L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20190712/1094748/1094748_17138345536257_big.jpg?w=1200",
+                name = "라이트웨이트 미들 삭Money(스 7팩 )[화이트]",
+                _price = Money(10_990),
+            ),
+            Product(
+                id = 34L,
+                imageUrl = "https://image.msscdn.net/thumbnails/images/goods_img/20210901/2101924/2101924_17138482498938_big.jpg?w=1200",
+                name = "플레인 더비 슈즈 [블랙]",
+                _price = Money(89_890),
+            ),
+        )
+}
